Revert "Fix bug"
