+++ frontend/src/components/MathjaxRenderer/index.js
@@ -5,7 +5,7 @@ import { Node, Context } from "../react-mathjax";
 import MathJaxConfig from "../../constants/MathJax_config";
 import dataSource from "../InputBox/dataSource";
 
-export default class MathjaxRenderer extends React.PureComponent {
+export default class MathjaxRenderer extends React.Component {
   _isMounted = false;
   state = {
     texCommandsInMathTag: [],
@@ -31,6 +31,13 @@ export default class MathjaxRenderer extends React.PureComponent {
 
   componentWillUnmount() {
     this._isMounted = false;
+  }
+
+  shouldComponentUpdate(nextProps, nextState) {
+    if (this.state.texCommandsInMathTag !== nextState.texCommandsInMathTag) {
+      return true;
+    }
+    return false;
   }
 
   regexMatch = (tex, note) => {
