+++ frontend/src/constants/ActionTypes.js
@@ -8,7 +8,6 @@ export const CLOSE_SUBS = "CLOSE_SUBS";
 export const LOAD_PARAS = "LOAD_PARAS";
 export const LOAD_PARAS_ERROR = "LOAD_PARAS_ERROR";
 export const PARA_ONCHANGE = "PARA_ONCHANGE";
-export const ADD_QUEUE = "ADD_QUEUE";
 export const CLEAR_QUEUE = "CLEAR_QUEUE";
 export const POP_QUEUE = "POP_QUEUE";
