+++ frontend/src/reducers/index.js
@@ -4,7 +4,6 @@ import subordinates from "./subordinates";
 import paras from "./paras";
 import currentpage from "./currentPage";
 import authentication from "./authenticate";
-import uploadingQueue from "./uploadQueue";
 
 export default combineReducers({
   windows,
@@ -12,5 +11,4 @@ export default combineReducers({
   paras,
   currentpage,
   authentication,
-  uploadingQueue,
 });
