--- frontend/src/utils/escapeString.js
@@ -1,4 +1,0 @@
-export function escapeString(str) {
-  // new line; tab; space;
-  return encodeURI(str);
-}
